polishing code
