+++ HTML/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <title>It works!</title>
+</head>
+<body>
+  <h1>It works!</h1>
+  <p>Serving index.html</p>
+</body>
+</html>
