+++ include/HTTPRequestLine.hpp
@@ -16,27 +16,31 @@
 #include <string>
 #include <sstream>
 
-/**
- * @brief Class responsible for parsing and validating HTTP request lines
- * 
- * The request line is the first line of an HTTP request and contains:
- * Method SP Request-URI SP HTTP-Version CRLF
- * Example: GET /index.html HTTP/1.1
- * 
- * This class handles:
- * - Parsing the method, path, and version from the request line
- * - Validation of each component according to HTTP specifications  
- * - Error detection and reporting
+/*
+  The request line is the first line of an HTTP request and contains:
+  [Method] SP [Request-URI] SP [HTTP-Version] CRLF
+  Example: GET /index.html HTTP/1.1
+  
+  This class handles:
+  - Parsing the method, path, and version from the request line
+  - Validation of each component according to HTTP specifications  
+  - Error detection and reporting
  */
 class HTTPRequestLine
 {
 private:
+    std::string _rawLine;   // Original request line for debugging
     std::string _method;    // HTTP Method (GET, POST, DELETE, etc.)
     std::string _path;      // Path of the requested resource
     std::string _version;   // HTTP Version (HTTP/1.0, HTTP/1.1)
-    std::string _rawLine;   // Original request line for debugging
     bool _isValid;          // Whether the request line is valid
     std::string _errorMessage; // Error message if parsing fails
+
+    // Helper methods (shuold be private because they are only used by the parseRequestLine method)
+    void setError(const std::string& message);
+    bool parseComponents(std::istringstream& iss);
+    void trimTrailingCR(std::string& line);
+};
 
 public:
     HTTPRequestLine();
@@ -65,12 +69,6 @@ public:
     
     // Reset the object to initial state
     void reset();
-    
-private:
-    // Helper methods
-    void setError(const std::string& message);
-    bool parseComponents(std::istringstream& iss);
-    void trimTrailingCR(std::string& line);
 };
 
 #endif
