+++ www/html/index.html
@@ -7,5 +7,6 @@
 <body>
   <h1>It works!</h1>
   <p>Serving index.html</p>
+  <button onclick="alert('Hello World!')">Click me</button>
 </body>
 </html>
